airflow/requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-postgres==5.11.3
flask<2.3
click<9
sqlalchemy<2
psycopg2-binary==2.9.9
python-gitlab==8.2.0
tree-sitter-language-pack>=0.13.0
qdrant-client>=1.16.2
fastembed>=0.8.0
tokenizers>=0.22.2
llama-index>=0.14.12
llama-index-embeddings-openai-like>=0.2.2
llama-index-llms-openai-like>=0.5.3
llama-index-postprocessor-jinaai-rerank>=0.5.0
llama-index-retrievers-bm25>=0.6.5
llama-index-vector-stores-qdrant>=0.9.1